# pi/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pi-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/terminal/connection-token": {
      "post": {
        "tags": [
          "edge"
        ],
        "summary": "Terminal Connection Token",
        "operationId": "terminal_connection_token_v1_terminal_connection_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalConnectionTokenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Terminal Connection Token V1 Terminal Connection Token Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/terminal/payment-intents": {
      "post": {
        "tags": [
          "edge"
        ],
        "summary": "Terminal Payment Intent",
        "operationId": "terminal_payment_intent_v1_terminal_payment_intents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPaymentIntentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Terminal Payment Intent V1 Terminal Payment Intents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/terminal/payment-intents/{payment_intent_id}": {
      "get": {
        "tags": [
          "edge"
        ],
        "summary": "Terminal Payment Intent Status",
        "operationId": "terminal_payment_intent_status_v1_terminal_payment_intents__payment_intent_id__get",
        "parameters": [
          {
            "name": "payment_intent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Intent Id"
            }
          },
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Terminal Payment Intent Status V1 Terminal Payment Intents  Payment Intent Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/registers/{cash_register_uuid}/display": {
      "get": {
        "tags": [
          "edge"
        ],
        "summary": "Get Register Display",
        "operationId": "get_register_display_v1_registers__cash_register_uuid__display_get",
        "parameters": [
          {
            "name": "cash_register_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cash Register Uuid"
            }
          },
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDisplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "edge"
        ],
        "summary": "Put Register Display",
        "operationId": "put_register_display_v1_registers__cash_register_uuid__display_put",
        "parameters": [
          {
            "name": "cash_register_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cash Register Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDisplayBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDisplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments": {
      "get": {
        "tags": [
          "edge"
        ],
        "summary": "List Payments",
        "operationId": "list_payments_v1_payments_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "waiter_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Waiter Uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments/{payment_id}/receipt": {
      "post": {
        "tags": [
          "edge"
        ],
        "summary": "Payment Receipt",
        "operationId": "payment_receipt_v1_payments__payment_id__receipt_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PaymentReceiptBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReceiptEscposResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments/{payment_id}/receipt/print": {
      "post": {
        "tags": [
          "edge"
        ],
        "summary": "Payment Receipt Print To Station",
        "operationId": "payment_receipt_print_to_station_v1_payments__payment_id__receipt_print_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentReceiptPrintBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReceiptPrintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscountIn": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "percent",
              "amount"
            ],
            "title": "Kind"
          },
          "value": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "value"
        ],
        "title": "DiscountIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineAdditionIn": {
        "properties": {
          "article_id": {
            "type": "integer",
            "title": "Article Id"
          },
          "qty": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Qty",
            "default": 1
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "article_id"
        ],
        "title": "LineAdditionIn"
      },
...
      "OrderLineIn": {
        "properties": {
          "article_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Article Id"
          },
          "qty": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Qty",
            "default": 1
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "additions": {
            "items": {
              "$ref": "#/components/schemas/LineAdditionIn"
            },
            "type": "array",
            "title": "Additions"
          },
          "discount": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "station_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Station Uuid"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "voucher_definition_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voucher Definition Uuid"
          },
          "voucher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voucher Name"
          },
          "value_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Cents"
          },
          "article_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Article Name"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "OrderLineIn"
      },
...
      "PaymentListItem": {
        "properties": {
          "payment_id": {
            "type": "integer",
            "title": "Payment Id"
          },
          "total_cents": {
            "type": "integer",
            "title": "Total Cents"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "payment_id",
          "total_cents",
          "item_count",
          "currency"
        ],
        "title": "PaymentListItem"
      },
      "PaymentReceiptBody": {
        "properties": {
          "reprint": {
            "type": "boolean",
            "title": "Reprint",
            "default": false
          },
          "paper_width": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "80mm",
                  "58mm",
                  "53mm"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Width"
          },
          "charset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pc858",
                  "pc850",
                  "cp1252",
                  "ascii"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Charset"
          }
        },
        "type": "object",
        "title": "PaymentReceiptBody"
      },
      "PaymentReceiptEscposResponse": {
        "properties": {
          "payment_id": {
            "type": "integer",
            "title": "Payment Id"
          },
          "escpos_payload": {
            "type": "string",
            "title": "Escpos Payload"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "escpos_payload"
        ],
        "title": "PaymentReceiptEscposResponse"
      },
      "PaymentReceiptPrintBody": {
        "properties": {
          "station_uuid": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Station Uuid"
          }
        },
        "type": "object",
        "required": [
          "station_uuid"
        ],
        "title": "PaymentReceiptPrintBody"
      },
      "PaymentReceiptPrintResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "print_job_id": {
            "type": "integer",
            "title": "Print Job Id"
          }
        },
        "type": "object",
        "required": [
          "print_job_id"
        ],
        "title": "PaymentReceiptPrintResponse"
      },
      "PaymentsListResponse": {
        "properties": {
          "payments": {
            "items": {
              "$ref": "#/components/schemas/PaymentListItem"
            },
            "type": "array",
            "title": "Payments"
          }
        },
        "type": "object",
        "required": [
          "payments"
        ],
        "title": "PaymentsListResponse"
      },
...
      "RegisterDisplayBody": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "payload": {
            "$ref": "#/components/schemas/RegisterDisplayPayload"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "RegisterDisplayBody"
      },
      "RegisterDisplayPayload": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "total_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cents"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineIn"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RegisterDisplayPayload"
      },
      "RegisterDisplayResponse": {
        "properties": {
          "cash_register_uuid": {
            "type": "string",
            "title": "Cash Register Uuid"
          },
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "payload": {
            "$ref": "#/components/schemas/RegisterDisplayPayload"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "cash_register_uuid",
          "event_id"
        ],
        "title": "RegisterDisplayResponse"
      },
...
      "TerminalConnectionTokenBody": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "TerminalConnectionTokenBody"
      },
      "TerminalPaymentIntentBody": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "amount_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "client_order_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Order Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "amount_cents"
        ],
        "title": "TerminalPaymentIntentBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}